requirements.txt
# LLM client (OpenAI-compatible, works with Gemini)
openai>=1.0.0

# Embeddings + re-ranking
sentence-transformers>=3.0.0

# Vector store with hybrid search
lancedb>=0.6.0
pyarrow>=14.0.0

# Data handling
pandas>=2.0.0
pdfplumber>=0.10.0

# Web scraping for auto-update
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Frontend
streamlit>=1.30.0

# Environment config
python-dotenv>=1.0.0

# Evaluation
scikit-learn>=1.3.0
